+++ libraries/matchers.rb
@@ -29,6 +29,7 @@ if defined?(ChefSpec)
     oneview_logical_enclosure:          [:update_from_group, :reconfigure, :set_script],
     oneview_logical_interconnect_group: standard_actions,
     oneview_logical_switch_group:       standard_actions,
+    oneview_logical_switch:             standard_actions + [:refresh],
     oneview_network_set:                standard_actions,
     oneview_rack:                       [:add, :remove, :add_if_missing, :add_to_rack, :remove_from_rack],
     oneview_server_hardware:            [:add, :add_if_missing, :remove, :refresh, :set_power_state, :update_ilo_firmware],