+++ spec/unit/resources/matchers_spec.rb
@@ -76,6 +76,12 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_logical_switch_group_if_missing('')
     expect(chef_run).to_not delete_oneview_logical_switch_group('')
 
+    # oneview_logical_switch
+    expect(chef_run).to_not create_oneview_logical_switch('')
+    expect(chef_run).to_not create_oneview_logical_switch_if_missing('')
+    expect(chef_run).to_not delete_oneview_logical_switch('')
+    expect(chef_run).to_not refresh_oneview_logical_switch('')
+
     # oneview_network_set
     expect(chef_run).to_not create_oneview_network_set('')
     expect(chef_run).to_not create_oneview_network_set_if_missing('')